+++ examples/super_simple.rs
@@ -1,6 +1,6 @@
 extern crate ggez;
 use ggez::*;
-use ggez::graphics::{DrawMode, Point};
+use ggez::graphics::{DrawMode, Point2};
 use std::time::Duration;
 
 struct MainState {
@@ -24,10 +24,7 @@ impl event::EventHandler for MainState {
         graphics::clear(ctx);
         graphics::circle(ctx,
                          DrawMode::Fill,
-                         Point {
-                             x: self.pos_x,
-                             y: 380.0,
-                         },
+                         Point2::new(self.pos_x, 380.0),
                          100.0,
                          2.0)?;
         graphics::present(ctx);